Fixed(?) depth testing

+++ nbody/Constants.h
@@ -4,8 +4,8 @@ const float step = 100.f; //86400.0f is 24 hours
 const double G = 6.673e-11;
 const double AU = 1.49597893e11;
 
-const float WIDTH = 800;
-const float HEIGHT = 800;
+const int WIDTH = 800;
+const int HEIGHT = 800;
 
 const int AMOUNT = 1000;
